flash/kernels: split into kernel_calls + kernel_opts, compose in kernels.py
flash_op/kernels imports flash/kernels which pulls in KernelControl at module
level — unavailable in the testing library (installed/test/<arch>/lib).

Split flash/kernels.py into three modules:
- kernel_calls.py: tensor prep + C++ op invocation via params-style V3 API.
  No KernelControl; safe to import from the testing library.
- kernel_opts.py: AttnOptionsWrapper + SdpaOpts mixin using KernelControl
  and attn_options. Tuning library only.
- kernels.py: composes SdpaOpts + kernel_calls classes into the original
  concrete kernel names; backward-compatible public API unchanged.

flash_op/kernels now imports SdpaCalls/attn_fwd/bwd_kernel_dk_dv from
kernel_calls, avoiding the KernelControl import entirely.

Also add from __future__ import annotations to tdesc.py to fix
  TypeError: unsupported operand type(s) for |: 'type' and 'str'
for the str | 'ImplSelector' annotation in the class body.
