+++ v3python/tune/tdesc.py
@@ -1,6 +1,8 @@
 # Copyright © 2025 Advanced Micro Devices, Inc.
 # SPDX-License-Identifier: MIT
 
+from __future__ import annotations
+
 import json
 import itertools
 from abc import ABC, abstractmethod
